requirements.txt
torch
pandas
numpy
transformers
accelerate
scikit-learn
matplotlib
pyyaml
